# Conflicts:
#	hustack-backend/judge0-client/src/main/java/vn/edu/hust/soict/judge0client/config/Judge0Config.java
#	hustack-backend/judge0-client/src/main/java/vn/edu/hust/soict/judge0client/utils/Judge0Utils.java

+++ hustack-backend/baseweb/src/main/java/com/hust/baseweb/applications/exam/service/impl/ExamQuestionServiceImpl.java
@@ -1,96 +1,76 @@
 package com.hust.baseweb.applications.exam.service.impl;
 
-
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hust.baseweb.applications.exam.entity.ExamQuestionEntity;
-import com.hust.baseweb.applications.exam.entity.ExamTestQuestionEntity;
+import com.hust.baseweb.applications.contentmanager.repo.MongoContentService;
+import com.hust.baseweb.applications.exam.entity.*;
 import com.hust.baseweb.applications.exam.model.ResponseData;
-import com.hust.baseweb.applications.exam.model.request.ExamQuestionDeleteReq;
-import com.hust.baseweb.applications.exam.model.request.ExamQuestionDetailsReq;
-import com.hust.baseweb.applications.exam.model.request.ExamQuestionFilterReq;
-import com.hust.baseweb.applications.exam.model.request.ExamQuestionSaveReq;
+import com.hust.baseweb.applications.exam.model.request.*;
+import com.hust.baseweb.applications.exam.model.response.ExamQuestionDetailsRes;
+import com.hust.baseweb.applications.exam.model.response.ExamQuestionFilterRes;
 import com.hust.baseweb.applications.exam.repository.ExamQuestionRepository;
+import com.hust.baseweb.applications.exam.repository.ExamQuestionTagRepository;
+import com.hust.baseweb.applications.exam.repository.ExamTagRepository;
 import com.hust.baseweb.applications.exam.repository.ExamTestQuestionRepository;
 import com.hust.baseweb.applications.exam.service.ExamQuestionService;
 import com.hust.baseweb.applications.exam.service.MongoFileService;
 import com.hust.baseweb.applications.exam.utils.DataUtils;
 import com.hust.baseweb.applications.exam.utils.SecurityUtils;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.Query;
-import jakarta.transaction.Transactional;
+import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.math.BigInteger;
+import jakarta.persistence.EntityManager;
+import jakarta.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
 @Slf4j
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class ExamQuestionServiceImpl implements ExamQuestionService {
 
-    private final ExamQuestionRepository examQuestionRepository;
-    private final ExamTestQuestionRepository examTestQuestionRepository;
-    private final MongoFileService mongoFileService;
-    private final ModelMapper modelMapper;
-    private final EntityManager entityManager;
-    private final ObjectMapper objectMapper;
+    ExamQuestionRepository examQuestionRepository;
+    ExamTagRepository examTagRepository;
+    ExamQuestionTagRepository examQuestionTagRepository;
+    ExamTestQuestionRepository examTestQuestionRepository;
+    MongoFileService mongoFileService;
+    ModelMapper modelMapper;
+    EntityManager entityManager;
+    ObjectMapper objectMapper;
 
     @Override
-    public Page<ExamQuestionEntity> filter(Pageable pageable, ExamQuestionFilterReq examQuestionFilterReq) {
-        StringBuilder sql = new StringBuilder();
-        sql.append("select\n" +
-                   "    *\n" +
-                   "from\n" +
-                   "    exam_question eq\n" +
-                   "where\n" +
-                   "    eq.created_by = :userLogin \n");
-        if(examQuestionFilterReq.getType() != null){
-            sql.append("and\n" +
-                       "    eq.type = :type \n");
-        }
-        if(DataUtils.stringIsNotNullOrEmpty(examQuestionFilterReq.getKeyword())){
-            sql.append("and\n" +
-                       "    ((lower(eq.code) like CONCAT('%', LOWER(:keyword),'%')) or \n" +
-                       "    (lower(eq.content) like CONCAT('%', LOWER(:keyword),'%'))) \n");
-        }
-        sql.append("order by created_at desc\n");
-
-        Query query = entityManager.createNativeQuery(sql.toString(), ExamQuestionEntity.class);
-        Query count = entityManager.createNativeQuery("select count(1) FROM (" + sql + ") as count");
-        query.setFirstResult(pageable.getPageNumber() * pageable.getPageSize());
-        query.setMaxResults(pageable.getPageSize());
-
-        query.setParameter("userLogin", SecurityUtils.getUserLogin());
-        count.setParameter("userLogin", SecurityUtils.getUserLogin());
-        if(examQuestionFilterReq.getType() != null){
-            query.setParameter("type", examQuestionFilterReq.getType());
-            count.setParameter("type", examQuestionFilterReq.getType());
-        }
-        if(DataUtils.stringIsNotNullOrEmpty(examQuestionFilterReq.getKeyword())){
-            query.setParameter("keyword", DataUtils.escapeSpecialCharacters(examQuestionFilterReq.getKeyword()));
-            count.setParameter("keyword", DataUtils.escapeSpecialCharacters(examQuestionFilterReq.getKeyword()));
-        }
-
-        long totalRecord = ((BigInteger) count.getSingleResult()).longValue();
-        List<ExamQuestionEntity> list = query.getResultList();
-        return new PageImpl<>(list, pageable, totalRecord);
+    public Page<ExamQuestionFilterRes> filter(Pageable pageable, ExamQuestionFilterReq examQuestionFilterReq) {
+        return examQuestionRepository.filter(
+            pageable,
+            SecurityUtils.getUserLogin(),
+            examQuestionFilterReq.getType(),
+            examQuestionFilterReq.getExamTagIds() != null ? List.of(examQuestionFilterReq.getExamTagIds().split(",")) : new ArrayList<>(),
+            examQuestionFilterReq.getLevel(),
+            examQuestionFilterReq.getExamSubjectId(),
+            examQuestionFilterReq.getKeyword()
+        );
+    }
+    private List<String> getIdInListExamTag(List<ExamTagEntity> examTags){
+        return examTags.stream()
+                       .map(ExamTagEntity::getId)
+                       .collect(Collectors.toList());
     }
 
     @Override
-    public ResponseData<ExamQuestionEntity> details(ExamQuestionDetailsReq examQuestionDetailsReq) {
-        ResponseData<ExamQuestionEntity> responseData = new ResponseData<>();
-        if(DataUtils.stringIsNotNullOrEmpty(examQuestionDetailsReq.getId())){
-            Optional<ExamQuestionEntity> examQuestionEntity = examQuestionRepository.findById(examQuestionDetailsReq.getId());
+    public ResponseData<ExamQuestionDetailsRes> details(String id) {
+        ResponseData<ExamQuestionDetailsRes> responseData = new ResponseData<>();
+        if(DataUtils.stringIsNotNullOrEmpty(id)){
+            Optional<ExamQuestionDetailsRes> examQuestionEntity = examQuestionRepository.findOneById(id);
             if(examQuestionEntity.isPresent()){
                 responseData.setHttpStatus(HttpStatus.OK);
                 responseData.setResultCode(HttpStatus.OK.value());
@@ -100,16 +80,16 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
             }
         }
 
-        if(DataUtils.stringIsNotNullOrEmpty(examQuestionDetailsReq.getCode())){
-            Optional<ExamQuestionEntity> examQuestionEntity = examQuestionRepository.findByCode(examQuestionDetailsReq.getCode());
-            if(examQuestionEntity.isPresent()){
-                responseData.setHttpStatus(HttpStatus.OK);
-                responseData.setResultCode(HttpStatus.OK.value());
-                responseData.setResultMsg("Success");
-                responseData.setData(examQuestionEntity.get());
-                return responseData;
-            }
-        }
+//        if(DataUtils.stringIsNotNullOrEmpty(code)){
+//            Optional<ExamQuestionDetailsRes> examQuestionEntity = examQuestionRepository.findOneByCode(code);
+//            if(examQuestionEntity.isPresent()){
+//                responseData.setHttpStatus(HttpStatus.OK);
+//                responseData.setResultCode(HttpStatus.OK.value());
+//                responseData.setResultMsg("Success");
+//                responseData.setData(examQuestionEntity.get());
+//                return responseData;
+//            }
+//        }
 
         responseData.setHttpStatus(HttpStatus.NOT_FOUND);
         responseData.setResultCode(HttpStatus.NOT_FOUND.value());
@@ -128,11 +108,23 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
             return responseData;
         }
 
-        List<String> filePaths = mongoFileService.storeFiles(files);
+        List<String> filePaths = new ArrayList<>();
+        if(files != null && files.length > 0){
+            filePaths = mongoFileService.storeFiles(files);
+        }
 
         ExamQuestionEntity examQuestionEntity = modelMapper.map(examQuestionSaveReq, ExamQuestionEntity.class);
         examQuestionEntity.setFilePath(String.join(";", filePaths));
-        examQuestionRepository.save(examQuestionEntity);
+        examQuestionEntity = examQuestionRepository.save(examQuestionEntity);
+
+        List<ExamQuestionTagEntity> examQuestionTagEntityList = new ArrayList<>();
+        for(ExamTagEntity item: examQuestionSaveReq.getExamTags()){
+            ExamQuestionTagEntity examQuestionTagEntity = new ExamQuestionTagEntity();
+            examQuestionTagEntity.setId(new ExamQuestionTagKey(item.getId(), examQuestionEntity.getId()));
+            examQuestionTagEntityList.add(examQuestionTagEntity);
+        }
+        examQuestionTagRepository.saveAll(examQuestionTagEntityList);
+
         responseData.setHttpStatus(HttpStatus.OK);
         responseData.setResultCode(HttpStatus.OK.value());
         responseData.setResultMsg("Thêm mới câu hỏi thành công");
@@ -152,7 +144,7 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
         }
 
         List<String> filePaths = new ArrayList<>();
-        if(files.length > 0){
+        if(files != null && files.length > 0){
             filePaths = mongoFileService.storeFiles(files);
         }
 
@@ -172,6 +164,25 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
             mongoFileService.deleteByPath(filePath);
         }
 
+        List<ExamTagEntity> examQuestionTagEntityExistList = examTagRepository.findAllByExamQuestionId(examQuestionEntity.getId());
+        List<ExamTagEntity> examQuestionTagEntityDeleteList = examQuestionTagEntityExistList.stream()
+                                                                                                    .filter(item -> !examQuestionSaveReq.getExamTags().contains(item))
+                                                                                                    .collect(Collectors.toList());
+        List<ExamTagEntity> examQuestionTagEntityAddList = examQuestionSaveReq.getExamTags().stream()
+                                                                                         .filter(item -> !examQuestionTagEntityExistList.contains(item))
+                                                                                         .collect(Collectors.toList());
+
+        for(ExamTagEntity item: examQuestionTagEntityDeleteList){
+            ExamQuestionTagEntity examQuestionTagEntity = new ExamQuestionTagEntity();
+            examQuestionTagEntity.setId(new ExamQuestionTagKey(item.getId(), examQuestionEntity.getId()));
+            examQuestionTagRepository.delete(examQuestionTagEntity);
+        }
+        for(ExamTagEntity item: examQuestionTagEntityAddList){
+            ExamQuestionTagEntity examQuestionTagEntity = new ExamQuestionTagEntity();
+            examQuestionTagEntity.setId(new ExamQuestionTagKey(item.getId(), examQuestionEntity.getId()));
+            examQuestionTagRepository.save(examQuestionTagEntity);
+        }
+
         responseData.setHttpStatus(HttpStatus.OK);
         responseData.setResultCode(HttpStatus.OK.value());
         responseData.setResultMsg("Cập nhật câu hỏi thành công");
@@ -180,9 +191,9 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
 
     @Override
     @Transactional
-    public ResponseData<ExamQuestionEntity> delete(ExamQuestionDeleteReq examQuestionDeleteReq) {
+    public ResponseData<ExamQuestionEntity> delete(String id) {
         ResponseData<ExamQuestionEntity> responseData = new ResponseData<>();
-        Optional<ExamQuestionEntity> examQuestionExist = examQuestionRepository.findById(examQuestionDeleteReq.getId());
+        Optional<ExamQuestionEntity> examQuestionExist = examQuestionRepository.findById(id);
         if(!examQuestionExist.isPresent()){
             responseData.setHttpStatus(HttpStatus.NOT_FOUND);
             responseData.setResultCode(HttpStatus.NOT_FOUND.value());
@@ -190,7 +201,7 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
             return responseData;
         }
 
-        List<ExamTestQuestionEntity> examTestQuestionEntityList = examTestQuestionRepository.findAllByExamQuestionId(examQuestionDeleteReq.getId());
+        List<ExamTestQuestionEntity> examTestQuestionEntityList = examTestQuestionRepository.findAllByExamQuestionId(id);
         if(!examTestQuestionEntityList.isEmpty()){
             responseData.setHttpStatus(HttpStatus.NOT_FOUND);
             responseData.setResultCode(HttpStatus.NOT_FOUND.value());
@@ -198,11 +209,18 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
             return responseData;
         }
 
-        String[] filePaths = examQuestionExist.get().getFilePath().split(";");
-        for(String filePath: filePaths){
-            if(DataUtils.stringIsNotNullOrEmpty(filePath)){
-                mongoFileService.deleteByPath(filePath);
+        if(examQuestionExist.get().getFilePath() != null){
+            String[] filePaths = examQuestionExist.get().getFilePath().split(";");
+            for(String filePath: filePaths){
+                if(DataUtils.stringIsNotNullOrEmpty(filePath)){
+                    mongoFileService.deleteByPath(filePath);
+                }
             }
+        }
+
+        List<ExamQuestionTagEntity> examQuestionTagEntityList = examQuestionTagRepository.findALLById_ExamQuestionId(id);
+        if(!examQuestionTagEntityList.isEmpty()){
+            examQuestionTagRepository.deleteAll(examQuestionTagEntityList);
         }
 
         examQuestionRepository.delete(examQuestionExist.get());
